{feat}: Fixing bug in search directory in server
The function would wait until the recursion base case to add anything, but we need the files in the none 'leaf node' directories as well. If we are in a directory that has other directories, we still make note of that to turn them into buttons in the file view. This will make navigation possible, as it was not with the previous version.